+++ server/src/routes/mercadoPago.js
@@ -1,6 +1,7 @@
 const mercadopago = require("mercadopago");
 require("dotenv").config();
 const { Router } = require("express");
+const Buys = require("../models/Buys");
 const router = Router();
 const access_token = "APP_USR-943877230059034-092202-de89def63c7dbdb08fd4752f4e60622d-1202964227";
 
@@ -14,7 +15,7 @@ mercadopago.configure({
       items: [
         {
           title: "Inscripcion Torneo",
-          unit_price: 1200,
+          unit_price: 1,
           quantity: 1,
         },
       ],
@@ -61,11 +62,11 @@ mercadopago.configure({
       },
     });
     //Aquí edito el status de mi orden
-    Order.findByPk(external_reference)
+    Buys.findByPk(external_reference)
       .then((order) => {
         order.payment_id = payment_id;
         order.payment_status = payment_status;
-        order.merchant_order_id = merchant_order_id;
+        // order.merchant_order_id = merchant_order_id;
         order.status = "completed";
         console.info("Salvando order");
         order
@@ -74,16 +75,16 @@ mercadopago.configure({
             console.info("redirect success");
             mensuAbonada.update({ abonado: true });
             mensuAbonada.save();
-            return res.redirect("http://localhost:3000/adminClinica");
+            return res.redirect("http://localhost:3000/home");
           })
           .catch((err) => {
             console.error("error al salvar", err);
-            return res.redirect("http://localhost:3000/adminClinica");
+            return res.redirect("http://localhost:3000/home");
           });
       })
       .catch((err) => {
         console.error("error al buscar", err);
-        return res.redirect("http://localhost:3000/adminClinica");
+        return res.redirect("http://localhost:3000");
       });
   
     //proceso los datos del pago
@@ -107,5 +108,58 @@ mercadopago.configure({
         });
       });
   });
+
+// mercadopago.configure({
+//   access_token: access_token
+// });
+
+// router.post('/', (req, res) => {
+//   /* aquí crea tu orden en la DB para el usuario logeado */
+//   const order = db.orders.create({ userId: req.userId, productId: req.body.productId }); // <--- pseudo-código
+
+//   Ahora le decimos a MP que cree la "preferencia". Asume que "order" tiene datos del producto
+//   mercadopago.preferences.create({
+//     external_reference: '1',
+//     notification_url: `http://localhost:3001/api/mercadopago/ipn`, // <-- apunta al endpoint definido en la línea 29
+//     items: [
+//       {
+//         title: "Inscripcion Torneo",
+//         unit_price: 1,
+//         quantity: 1,
+//       }
+//     ]
+//   }).then((preference) => {
+//     res.json({ preferenceId: preference.id });
+//   });
+// });
+
+// router.get('/api/mercadopago/ipn', (req, res) => {
+//   if (req.params.type === 'payment') { // hay otros, nos importa solo payment
+//     const paymentId = req.params.data.id; // ID de payment en MercadoPago
+    
+//     Documentación de pagos: https://www.mercadopago.cl/developers/es/reference/payments/_payments_search/get/
+//     mercadopago.payment.get(paymentId).then((error, payment) => {
+//       Obtenemos los datos del pago desde MP
+//       const orderId = payment.external_reference; // esto es el ID de la orden que especificamos en la linea 15
+      
+//       buscamos en nuestra db la orden
+//       db.orders.find(orderId).then((order) => {
+
+//         if (order.totalPrice === payment.transaction_amount) { // para que no se nos hagan los vivos XD
+//           order.status = payment.status; // hay muchos estados, revisa https://www.mercadopago.cl/developers/es/reference/payments/_payments_search/get/
+          
+//           comprobamos que sea "approved" y que no hayamos entregado ya el pedido... recuerda que "order" es algo que
+//           debes implementar tu, que podría tener un cambpo "delivered" para saber si ya hiciste entrega o no del
+//           pedido
+//           if (order.status === 'approved' && !order.delivered) {
+//             deliverOrder(order); // función ficticia que debes implementar... es básicamente "entregar" el producto
+//           }
+//         }
+        
+//       });
+//     });
+//   }
+// });
+
   
   module.exports = router;
